endpoint & service for collecting email addresses to send contribution reminders
