+++ support-frontend/app/controllers/ReminderController.scala
@@ -6,20 +6,19 @@ import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
 import play.api.libs.circe.Circe
 import play.api.mvc.{AbstractController, Action, ControllerComponents}
-import codecs.CirceDecoders._
-import services.SendReminderEmailService
+import services.RemindMeService
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 
 class ReminderController(components: ControllerComponents,
                          actionRefiners: CustomActionBuilders,
-                         sendReminderEmail: SendReminderEmailService)
+                         sendReminderEmail: RemindMeService)
                         (implicit ec: ExecutionContext)
   extends AbstractController(components) with Circe {
 
   import actionRefiners._
 
-  def reminderEvent(): Action[ReminderEventRequest] = PrivateAction.async(circe.json[ReminderEventRequest]) { implicit request =>
+  def sendReminderCreatedEvent(): Action[ReminderEventRequest] = PrivateAction.async(circe.json[ReminderEventRequest]) { implicit request =>
 
     val result = sendReminderEmail(request.body.email)
     result.map { res =>
