+++ support-frontend/conf/routes
@@ -66,7 +66,7 @@ GET  /support-workers/status                                       controllers.S
 # the need for a client token
 POST  /contribute/send-marketing                                   controllers.IdentityController.submitMarketing
 
-POST  /contribute/remind-me                                        controllers.ReminderController.reminderEvent
+POST  /contribute/remind-me                                        controllers.ReminderController.sendReminderCreatedEvent
 
 # ------ Identity ------ #
 PUT /identity/set-password-guest                                   controllers.IdentityController.setPasswordGuest
