+++ support-frontend/{test}
@@ -1,0 +1,36 @@
+HTTP/1.1 415 Unsupported Media Type
+Server: nginx/1.17.3
+Date: Wed, 09 Oct 2019 10:45:46 GMT
+Content-Type: text/html; charset=UTF-8
+Content-Length: 1187
+Connection: keep-alive
+X-Powered-By: Express
+vary: DNT
+expires: Wed, 09 Oct 2019 10:46:16 GMT
+cache-control: max-age=30, stale-while-revalidate=3, stale-if-error=864000
+referrer-policy: origin-when-cross-origin, strict-origin-when-cross-origin
+x-frame-options: DENY
+x-xss-protection: 1; mode=block
+surrogate-control: max-age=30, stale-while-revalidate=3, stale-if-error=864000
+x-content-type-options: nosniff
+content-security-policy: default-src 'self' http://localhost:9000 members-data-api.thegulocal.com metric-push-api-code.support.guardianapis.com www.paypalobjects.com www.paypal.com www.sandbox.paypal.com js.stripe.com pasteup.guim.co.uk ophan.theguardian.com j.ophan.co.uk media.guim.co.uk uploads.guim.co.uk www.google-analytics.com www.googletagmanager.com tagmanager.google.com polyfill.io www.googleadservices.com googleads.g.doubleclick.net www.google.com www.google.co.uk optimize.google.com static.ads-twitter.com bat.bing.com bid.g.doubleclick.net t.co analytics.twitter.com stats.g.doubleclick.net www.youtube-nocookie.com connect.facebook.net www.facebook.com consumer.krxd.net cdn.krxd.net beacon.krxd.net secure.adnxs.com checkout.stripe.com fonts.googleapis.com ssl.gstatic.com www.gstatic.com fonts.gstatic.com sentry.io www.dwin1.com data: wss: 'unsafe-inline' 'unsafe-eval'  q.stripe.com payment.code.dev-guardianapis.com https://interactive.guim.co.uk/ https://www.theguardian.com/ https://theguardian.com/
+x-permitted-cross-domain-policies: master-only
+
+HTTP/1.1 415 Unsupported Media Type
+Server: nginx/1.17.3
+Date: Wed, 09 Oct 2019 10:45:46 GMT
+Content-Type: text/html; charset=UTF-8
+Content-Length: 1187
+Connection: keep-alive
+X-Powered-By: Express
+vary: DNT
+expires: Wed, 09 Oct 2019 10:46:16 GMT
+cache-control: max-age=30, stale-while-revalidate=3, stale-if-error=864000
+referrer-policy: origin-when-cross-origin, strict-origin-when-cross-origin
+x-frame-options: DENY
+x-xss-protection: 1; mode=block
+surrogate-control: max-age=30, stale-while-revalidate=3, stale-if-error=864000
+x-content-type-options: nosniff
+content-security-policy: default-src 'self' http://localhost:9000 members-data-api.thegulocal.com metric-push-api-code.support.guardianapis.com www.paypalobjects.com www.paypal.com www.sandbox.paypal.com js.stripe.com pasteup.guim.co.uk ophan.theguardian.com j.ophan.co.uk media.guim.co.uk uploads.guim.co.uk www.google-analytics.com www.googletagmanager.com tagmanager.google.com polyfill.io www.googleadservices.com googleads.g.doubleclick.net www.google.com www.google.co.uk optimize.google.com static.ads-twitter.com bat.bing.com bid.g.doubleclick.net t.co analytics.twitter.com stats.g.doubleclick.net www.youtube-nocookie.com connect.facebook.net www.facebook.com consumer.krxd.net cdn.krxd.net beacon.krxd.net secure.adnxs.com checkout.stripe.com fonts.googleapis.com ssl.gstatic.com www.gstatic.com fonts.gstatic.com sentry.io www.dwin1.com data: wss: 'unsafe-inline' 'unsafe-eval'  q.stripe.com payment.code.dev-guardianapis.com https://interactive.guim.co.uk/ https://www.theguardian.com/ https://theguardian.com/
+x-permitted-cross-domain-policies: master-only
+
